--- asm/Z2AudioLib/Z2SoundInfo/getCategory__11Z2SoundInfoCF10JAISoundID.s
@@ -1,3 +1,0 @@
-lbl_802BB0D8:
-/* 802BB0D8  88 64 00 01 */	lbz r3, 1(r4)
-/* 802BB0DC  4E 80 00 20 */	blr
