--- asm/Z2AudioLib/Z2SoundInfo/getSeqInfo__11Z2SoundInfoCF10JAISoundIDP6JAISeq.s
@@ -1,12 +1,0 @@
-lbl_802BB8B4:
-/* 802BB8B4  94 21 FF F0 */	stwu r1, -0x10(r1)
-/* 802BB8B8  7C 08 02 A6 */	mflr r0
-/* 802BB8BC  90 01 00 14 */	stw r0, 0x14(r1)
-/* 802BB8C0  80 04 00 00 */	lwz r0, 0(r4)
-/* 802BB8C4  90 01 00 08 */	stw r0, 8(r1)
-/* 802BB8C8  38 81 00 08 */	addi r4, r1, 8
-/* 802BB8CC  48 00 02 7D */	bl getSoundInfo___11Z2SoundInfoCF10JAISoundIDP8JAISound
-/* 802BB8D0  80 01 00 14 */	lwz r0, 0x14(r1)
-/* 802BB8D4  7C 08 03 A6 */	mtlr r0
-/* 802BB8D8  38 21 00 10 */	addi r1, r1, 0x10
-/* 802BB8DC  4E 80 00 20 */	blr
