+++ include/JSystem/JAudio2/JAIStreamDataMgr.h
@@ -4,7 +4,7 @@
 #include "JSystem/JAudio2/JAISound.h"
 
 struct JAIStreamDataMgr {
-    virtual void getStreamFileEntry(JAISoundID) = 0;
+    virtual s32 getStreamFileEntry(JAISoundID) = 0;
     virtual ~JAIStreamDataMgr();
 };
